+++ src/client/src/pages/overview/boreholeTable/boreholeTable.tsx
@@ -40,6 +40,19 @@ export interface BoreholeTableProps {
   isBusy: boolean;
 }
 
+const formatWithThousandSeparator = (value: number | null): string => {
+  if (value == null) return "-";
+
+  // Format number using de-CH
+  const formatted = new Intl.NumberFormat("de-CH", {
+    useGrouping: true,
+    minimumFractionDigits: 2,
+  }).format(value);
+
+  // Ensure thousand separators are always a standard single quote (')
+  return formatted.replace(/’/g, "'");
+};
+
 export const BoreholeTable: FC<BoreholeTableProps> = ({
   boreholes,
   paginationModel,
@@ -157,9 +170,7 @@ export const BoreholeTable: FC<BoreholeTableProps> = ({
     },
     {
       field: "total_depth",
-      valueGetter: value => {
-        return Math.round(value * 100) / 100;
-      },
+      valueGetter: (value, row) => formatWithThousandSeparator(Math.round(row.total_depth * 100) / 100),
       headerName: t("totaldepth"),
       flex: 1,
     },
@@ -177,25 +188,19 @@ export const BoreholeTable: FC<BoreholeTableProps> = ({
     },
     {
       field: "reference_elevation",
-      valueGetter: value => {
-        return Math.round(value * 100) / 100;
-      },
+      valueGetter: (value, row) => formatWithThousandSeparator(Math.round(row.reference_elevation * 100) / 100),
       headerName: t("reference_elevation"),
       flex: 1,
     },
     {
       field: "location_x",
-      valueGetter: (value, row) => {
-        return `${Math.round(row.location_x * 100) / 100 || "-"}`;
-      },
+      valueGetter: (value, row) => formatWithThousandSeparator(Math.round(row.location_x * 100) / 100),
       headerName: t("location_x"),
       flex: 1,
     },
     {
       field: "location_y",
-      valueGetter: (value, row) => {
-        return `${Math.round(row.location_y * 100) / 100 || "-"}`;
-      },
+      valueGetter: (value, row) => formatWithThousandSeparator(Math.round(row.location_y * 100) / 100),
       headerName: t("location_y"),
       flex: 1,
     },